peter-voice/requirements.txt
# Peter Voice Desktop Client - Dependencies
# Requires Python 3.13 (kokoro-onnx doesn't support 3.14)

# TTS - Kokoro ONNX
kokoro-onnx>=0.4.0

# STT - Moonshine
useful-moonshine-onnx>=1.0.0

# Audio I/O
sounddevice>=0.5.0
soundfile>=0.12.0
numpy>=1.26.0

# VAD - Voice Activity Detection
torch>=2.0.0
silero-vad>=5.1

# System tray
pystray>=0.19.0
Pillow>=10.0.0

# Global hotkeys
pynput>=1.7.0

# HTTP client
aiohttp>=3.9.0

# Config
python-dotenv>=1.0.0
